requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-api-python-client==2.118.0
numpy==1.26.3 